test: Replace test_num_dependencies with github action
The purpose of the test_num_dependencies test was to be a "prompt" about
dependency changes in PRs. However, manually updating the
dependencies.txt file is incompatible with letting dependabot manage
automated dependency updates (#3585). We therefore decided to replace
the test with a different "prompt" in the form of an optional GitHub
Actions check that fails whenever a lockfile is touched.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>