test: extract the memory monitor from Microvm class
The memory monitor is a subsystem that monitors Firecracker's RSS
doesn't go above 5MB. It is implemented as a thread that shells out to
pmap to get the RSS every 50ms. It is intrussive and brittle as it
sometimes break when running tests in parallel.

Make the monitor opt-in, though currently no test seems to directly need
it.

Signed-off-by: Pablo Barbáchano <pablob@amazon.com>