test: move firecracker compilation to fixture setup
Right now the first test "call" phase takes longer because it includes
compiling Firecracker. By moving it to the factory it will show up in
the "setup" phase, so the test timings will be more accurate.

While at it, clean a bit the interface by moving the optional parameters
to factory defaults.

Signed-off-by: Pablo Barbáchano <pablob@amazon.com>