refactor: Change `GuestMemoryExtension::with_file` into `memfd_backed`
The function was only ever called with a `File` object created from a
`memfd`. Thus instead of the call-site have to do two function calls
(one for the memfd and then another for creating the guest memory mmap,
which because of this had to do a `fstat` call to determine how big the
memfd is), have the `GuestMemoryMmap` factory directly create the
`memfd`.

This is mainly preparatory to prevent callers to pass mismatched
configurations to `memfd_create` and the subsequent `from_file` call
(e.g. different huge page configs).

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>