fix: fsync metrics in Metrics::write
We call `Metrics::write` from the context of a signal handler that
terminates the process via `_exit`. This means that data written via
`write` that not been synced to disk yet will be lost (as only `exit`
flushes I/O buffers). Thus, manually ensure `fsync` is called on the
metrics file.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>