refactor: Do not store *const Vcpu in TLS
We only need TLS so that we can set `kvm_run.immediate_exit` to 1 in a
signal handler (for when we want to pause a vcpu and need to get KVM to
exit as soon as possible). Thus, only store the `KvmRunWrapper` here.
This is because aliasing the entire Vcpu object is definitely not okay
from Rust's safety model, while aliasing a mmap area that's shared with
the kernel is more of a gray area.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>