test: fail test if firecracker was unexpectedly killed by signal
The comment about "it being too late to assert" was outdated, as it only
applied back when `kill` was called inside of destructors, and before we
configured pytest to stop ignoring exceptions in destructors.

Nowadays, we call `kill` inside of test teardown, were it is very
possible indeed to raise exceptions. So do that.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>