refactor(queue): update `is_valid` check
Previously `is_valid` was doing 2 things:
- call `is_layout_valid`
- check the number of descriptors in the queue
is less than the queue size

`is_valid` was called before the device activation.

With new approach for accessing queue objects, we
cannot call `is_valid` before device activation, because
we need pointers in the `Queue` to be set in order to
calculate the `len` of the queue.

Considerring all of this we can remove `is_valid` and
just use `is_layout_valid` instead.
The check for valid queue size can be move into `initialize`
method.

Signed-off-by: Egor Lazarchuk <yegorlz@amazon.co.uk>