test: stop logging warnings in run_cmd if exception is thrown
If the exception ends up being fatal, then it will get printed in full
anyway, and it contains the entire log message anyway, and we just end
up with duplicate messages. If its not fatal (e.g. caught), then its
arguably up to the caller to decide whether it should be logged.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>