feat: allocate a separate memslot for DMA
On ARM, we can use FDT to describe "reserved-memory", and force device
drivers to use these reserved memory ranges for all data that the device
needs to access (e.g. vrings and io buffers). Inside the guest,
swiotlb/DMA APIs will be used to either directly allocate data
structures into this region, or to bounce userspace I/O buffers through
this region. Note that the guest kernel needs to be compiled with
`CONFIG_DMA_RESTRICTED_POOL=y` for this.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>