refactor: hoist vm creation out off create_vmm_and_vcpus
When creating guest_memfd-backed memory, we need to create a
GuestMemoryMmap from a guest_memfd. However, we can only create
guest_memfd's from a VmFd, meaning it needs to happen between creation
of the struct Vm, and the struct Vmm. This would place it inside
create_vmm_and_vcpus, but in this function we don't have any of the
information needed to construct the gmem (secret freedom enabled?
size?). To resolve this, move vm creation outside of this function, so
that we can create the guest_memfd in build_microvm_for_boot, before
calling create_vmm_and_vcpus.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>