fix: use BufReader for snapshot restore
In commit e7504aeb77b8 ("refactor: cleanup vmm::snapshot module"),
firecracker started reading the snapshot vmstate file in a single pass
instead of first loading it into a Vec and then deserializing. This
seems to have caused some performance regression due to the deserializer
doing many successive reads, resulting in many read(2) syscalls.
Counteract this with a `BufReader`.

Since a BufReader might got compute-ahead of the CRC, we need to use
some math: CRC is designed such that `crc(crc(0, bytes), crc(0,
bytes).to_ne_bytes()) == 0`. So it doesnt matter if we readahead the
CRC, because if the CRC of the snapshot matches the CRC stored in the
file, then (and only then) the CRC of the entire file (e.g. snapshot
data + stored CRC) will be 0.

Switch over to bincode::decode_from_reader, which has an optimize
implementation for slices and BufReaders (which avoids the performance
penalty incurred by `decode_from_std_read` due to repeated calls to
.read(), even if a BufReader is used).

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>